+++ src/Localization/Localization.mjs
@@ -1,3 +1,3 @@
 /**
- * @typedef {{getLanguage: (module: string) => Promise<{language: string}>, translate: (module: string, key: string, placeholders?: {[key: string]: string} | null) => Promise<string>}} Localization
+ * @typedef {{getLanguage: () => Promise<{language: string}>, translate: (module: string, key: string, placeholders?: {[key: string]: string} | null) => Promise<string>}} Localization
  */