+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2024-04-16-1
+
+Changes:
+
+- Move module to texts
+
 ## v2024-04-15-1
 
 Changes: